ctl: rename to go-gettable kubectl-cert_manager
People can be using:

go get github.com/jetstack/cert-manager/cmd/kubectl-cert_manager@latest
