--- cmd/ctl/pkg/convert/BUILD.bazel
@@ -1,38 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["convert.go"],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/convert",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/ctl:go_default_library",
-        "//pkg/logs:go_default_library",
-        "@com_github_spf13_cobra//:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/internalversion:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime/serializer:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime/serializer/json:go_default_library",
-        "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
-        "@io_k8s_cli_runtime//pkg/printers:go_default_library",
-        "@io_k8s_cli_runtime//pkg/resource:go_default_library",
-        "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
-        "@io_k8s_kubectl//pkg/util/i18n:go_default_library",
-        "@io_k8s_kubectl//pkg/util/templates:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
