--- cmd/ctl/pkg/create/BUILD.bazel
@@ -1,31 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["create.go"],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/create",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//cmd/ctl/pkg/create/certificaterequest:go_default_library",
-        "@com_github_spf13_cobra//:go_default_library",
-        "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
-        "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//cmd/ctl/pkg/create/certificaterequest:all-srcs",
-    ],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
