--- cmd/ctl/pkg/renew/BUILD.bazel
@@ -1,44 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["renew.go"],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/renew",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/api/util:go_default_library",
-        "//pkg/apis/certmanager/v1:go_default_library",
-        "//pkg/apis/meta/v1:go_default_library",
-        "//pkg/client/clientset/versioned:go_default_library",
-        "@com_github_spf13_cobra//:go_default_library",
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-        "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
-        "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//rest:go_default_library",
-        "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
-        "@io_k8s_kubectl//pkg/util/i18n:go_default_library",
-        "@io_k8s_kubectl//pkg/util/templates:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["renew_test.go"],
-    embed = [":go_default_library"],
-    deps = ["@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library"],
-)
