--- cmd/ctl/pkg/status/certificate/BUILD.bazel
@@ -1,60 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "certificate.go",
-        "types.go",
-    ],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/status/certificate",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//cmd/ctl/pkg/status/util:go_default_library",
-        "//pkg/apis/acme/v1beta1:go_default_library",
-        "//pkg/apis/certmanager/v1:go_default_library",
-        "//pkg/client/clientset/versioned:go_default_library",
-        "//pkg/ctl:go_default_library",
-        "//pkg/util/pki:go_default_library",
-        "//pkg/util/predicate:go_default_library",
-        "@com_github_spf13_cobra//:go_default_library",
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-        "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
-        "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//rest:go_default_library",
-        "@io_k8s_client_go//tools/reference:go_default_library",
-        "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
-        "@io_k8s_kubectl//pkg/describe:go_default_library",
-        "@io_k8s_kubectl//pkg/util/i18n:go_default_library",
-        "@io_k8s_kubectl//pkg/util/templates:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["certificate_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//pkg/apis/acme/v1beta1:go_default_library",
-        "//pkg/apis/certmanager/v1:go_default_library",
-        "//pkg/apis/meta/v1:go_default_library",
-        "//test/unit/gen:go_default_library",
-        "@com_github_stretchr_testify//assert:go_default_library",
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-    ],
-)
