--- cmd/ctl/pkg/status/util/BUILD.bazel
@@ -1,29 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["util.go"],
-    importpath = "github.com/jetstack/cert-manager/cmd/ctl/pkg/status/util",
-    visibility = ["//visibility:public"],
-    deps = [
-        "@io_k8s_api//core/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
-        "@io_k8s_apimachinery//pkg/util/duration:go_default_library",
-        "@io_k8s_kubectl//pkg/describe:go_default_library",
-        "@io_k8s_kubectl//pkg/util/event:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
