+++ cmd/kubectl-cert_manager/cmd/BUILD.bazel
@@ -1,0 +1,35 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["cmd.go"],
+    importpath = "github.com/jetstack/cert-manager/cmd/kubectl-cert_manager/cmd",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//cmd/ctl/pkg/convert:go_default_library",
+        "//cmd/ctl/pkg/create:go_default_library",
+        "//cmd/ctl/pkg/inspect:go_default_library",
+        "//cmd/ctl/pkg/renew:go_default_library",
+        "//cmd/ctl/pkg/status:go_default_library",
+        "//cmd/ctl/pkg/version:go_default_library",
+        "@com_github_spf13_cobra//:go_default_library",
+        "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
+        "@io_k8s_client_go//plugin/pkg/client/auth:go_default_library",
+        "@io_k8s_klog_v2//:go_default_library",
+        "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
+    ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
