+++ cmd/kubectl-cert_manager/pkg/inspect/secret/secret.go
